double as the output type instead of float is introduced

+++ inc/OperatorHeaders/bta.h
@@ -43,9 +43,9 @@ virtual float calc(const float &number1, const int &number2);
 virtual float calc(const int &number1, const float &number2);
 virtual float calc(const float &a, const float &b);
 
-virtual float calc(const double &a, const double &b);
-virtual float calc(const double &a, const float &b);
-virtual float calc(const float &a, const double &b);
+virtual double calc(const double &a, const double &b);
+virtual double calc(const double &a, const float &b);
+virtual double calc(const float &a, const double &b);
 
 
 virtual int calc_ref(const int &a, const int &b);
