+++ inc/OperatorHeaders/hw_ac.h
@@ -32,15 +32,15 @@ class hw_ac {
 		//void tbl_gen(const size_t &msb, const size_t &lsb, int (*calc) (const int &, const int &), int (*calc_ref) (const int &, const int &));
 		virtual void tbl_gen()=0;
 
-		virtual float calc(const double &a, const float &b) = 0; // pure virtual
-		virtual float calc(const float &a, const double &b) = 0; // pure virtual
+		virtual double calc(const double &a, const float &b) = 0; // pure virtual
+		virtual double calc(const float &a, const double &b) = 0; // pure virtual
         virtual int calc(const int &a, const int &b) = 0; // pure virtual
 		virtual int calc(const long &a, const int &b) = 0; // pure virtual
 		virtual int calc(const long &a, const long &b) = 0; // pure virtual
         virtual int calc(const int &a, const long &b) = 0; // pure virtual
         virtual unsigned int calc(const unsigned int &a, const unsigned int &b) = 0; // pure virtual
         
-		virtual float calc(const double &a, const double &b) = 0; // pure virtual
+		virtual double calc(const double &a, const double &b) = 0; // pure virtual
         
         virtual size_t get_ianum_bits(void)=0;
         virtual int calc(const unsigned int &a, const int &b) = 0; // pure virtual
